feat: implement app.ts for supertest
